NP-9276 Move files
