+++ src/pages/public_registration/public_files/preview_file/PreviewOfficeFile.tsx
@@ -1,5 +1,5 @@
 import { Box } from '@mui/material';
-import { dataTestId } from '../../../utils/dataTestIds';
+import { dataTestId } from '../../../../utils/dataTestIds';
 import { CommonPreviewProps } from './PreviewFile';
 
 interface PreviewOfficeFileProps extends CommonPreviewProps {
